Moved build_graph contents to other files Made integrate its own module with a CLI-ready function at top
